[NG] Checkbox
Angular component for checkboxes, mostly useful when you don't have an
easily available ID to bind the label to the checkbox. Our component
offers its own simple synchronous two-way binding, but also supports
ngModel with all its features.
This component is here as an internal utility for now, but it is solid
and tested enough to be shipped publicly. We just need documentation.

Also includes a small clean-up of checkboxes CSS to be more resilient
to positioning and size modifications.

Signed-off-by: Eudes Petonnet-Vincent <petonnet.eudes@gmail.com>