[NG] Datagrid
First version of the Datagrid Angular component.
It comes with a first set of features, including sorting, filtering, pagination and
multi-selection. It supports pre-loaded data as well as server-driven data, where
only the currently displayed items are loaded in the browser. It doesn’t contain any
of the non-critical features we have in our backlog, these will come in later
versions.

It uses a declarative interface, meaning big configuration JSONs are a no-no.
Instead we play with dependency injection and child querying to activate features
and communicate between components.

Style-wise, it mostly reuses the styles from static tables, with a few customizations.

And unit tests. Lots and lots and lots of unit tests.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>