[NG] Fix Datagrid's undefined trackBy
Angular 2.4.1 introduced an involuntary breaking change. It will be rollbacked to a simple warning in 2.4.2, but will be re-introduced in the next major version, so we need to fix it anyway. See https://github.com/angular/angular/issues/13641 for more details.
This commit explicitly states the default "identity" trackBy function for datagrid, to avoid future issues.

It also upgrades our project to the latest version of Angular, and fixes an issue with Jasmine typings being out of sync with the actual Jasmine version (which broke one of the unit tests).

Fixes #282, #305.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>