[NG] Support manual selection of Datagrid row
Binding to [clrDgSelected] was completely ignored when the entire datagrid was not selectable, which prevented people from manually creating single-row selection with the entire row as click target. They had to bind to an internal class, and even then it resulted in conflicts with our own implementation overriding their initial value.

To fix this, we now support the [clrDgSelected] input on rows all the time, but it only handles styling when the whole datagrid is not selectable.

Fixes #388.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>