[NG] Datagrid: expose a method to allow manual refresh
When consumers mutate the data they pass to the Datagrid, we have no way of detecting that the data has changed and that it should be re-processed by the filters, the comparators, ...
Although we recommend the data given to Datagrid should be immutable, we now provide a way for consumers to manually re-trigger this processing if they still decide to mutate the data. The way to do this is simply to grab the instance the the Datagrid (through @ViewChild, template variables, ...) and to call `.dataChanged()` on it.

Fixes #140 (kinda).

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>