[NG] Fix datagrid pagination page values when datagrid is empty
When 0 items in the datagrid, and paging active, firstItem was reported as 1, and lastItem as the page size. This could lead to footers like "1 - 10 of 0 items". It is correct to return -1 in this case, as these value are indices, and 0 would be valid index of a non-existing 1st element (thus "1 - 1 of 0").

Fixes #518

Signed-off-by: Ivan Donchev <idonchev@vmware.com>