[NG] Allow reverse order in datagrid 2-way binding sort attribute
- Introduced an exported const Enum `SortOrder` that defines the 3 possible
sorting states for a datagrid column: Unsorted, Asc and Desc.
- Removed clrDgSorted:boolean 2-way binding
- Introduced clrDgSortOrder: SortOrder 2-way binding:
  - replaced the existing logic around the sorted state (2 states: ON or OFF)
  with a tri-state sorting logic (following the 3 SortOrder enum values)
  - made sure all existing behaviour is preserved:
    - the datagrid-column reacts to the `sort` provider change events, i.e. when
    sort is cleared or toggled
    - modified the public sort() method to allow forcing the reverse sorting
- Modified the Sort provider toggle method to allow forcing the reverse sorting. It
is an optional parameter, so it does not break backward compatibility within
clarity-angular implementation
- modified and added some new test cases for the datagrid-column component and the
Sort provider

I left the original boolean `clrDgSorted` 2-way binding in clr-dg-column to allow
backward compatibility for the component consumption. I added some deprecation
annotations around all code that needs to be removed in the next major release.

/!\ Using both 2-way bindings at the same time will have unpredicted results, so
the using code needs to either use `clrDgSorted` or `clrDgSortOrder`. This needs
be explicitly mentioned in the doc.

This closes #586

Signed-off-by: Clement Cureau <clement.cureau@gmail.com>