[NG] Fixing alt cancel issue in new wizard (#708)
These changes Address the following issues with the 0.9.0 wizard:

• an issue where and alt cancel routine was fired twice at the wizard level
• an issue where a page level alt cancel routine fired and was followed by a wizard level alt cancel
• an issue with enter and leave animations in the modal

To resolve this issue, I gave precedence to page level alt cancel routines. If the current page has defined either and preventDefault or stopCancel, it will prevent comparable events at the wizard level from firing. This was already the case for finish, previous, next, and danger buttons. The logic was not implemented for cancel buttons, however.

Tested in:
✔︎ Chrome

Signed-off-by: Scott Mathis <smathis@vmware.com>