[ICONS] adding standard size icons for alert icons
Icons that are used in alerts are visually smaller than the other icons because they need to have thicker strokes and space around them to accommodate that.

These standard icons allow the usual alert type of icons next to other icons in the library.

The change was made to be as unobtrusive as possible. I did not make an value judgements on which style of icon should take precedence.

Closes: #701

Tested in:
✔︎ Chrome

Signed-off-by: Scott Mathis <smathis@vmware.com>

[ICONS] updated gemini tests

The hard way…

Signed-off-by: Scott Mathis <smathis@vmware.com>