[NG] Popovers refactor
Dropdown, tooltips and signposts have been moved to a consistent
Popover implementation, that plays better with Angular's component
lifecycle. They all support the use of `*clrIfOpen` on the popover,
but always make it optional (but recommended).

It includes a non-breaking renaming of `clrDropdownToggle` to
`clrDropdownTrigger`, once again for consistency.

BREAKING CHANGE: In all 3 cases, the positioning and sizing inputs of the popover
have been moved to the popover itself: dropdown-menu, tooltip-content
and signpost-content.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>