[NG] Datagrid chocolate
Cleaning up some lifecycle, adding a generic solution to projected
siblings updating each other or template children updating the parent
view when created.

We will have to be very careful, any future commit using or modifying
any file in the chocolate factory should be properly monitored
and evaluated. The valid use cases for it are rare, it turns out the
hideable columns didn't even need it, it was just a case of messed
up lifecycle where everything happened after the view was already
initialized. Beware! The chocolate factory is much scarier than the
name indicates. Well, not as scary as Johnny Depp's Willy Wonka, but
still scary.

Fixes #623, #744, #817, #862, #928, #950 and #951
(not counting duplicates).

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>