[NG] allow action overflow to be tabbable and refactor to use abstract popover
Change the action overflow icon to a button, so when using the tab key you can navigate for accessibility. Refactoring the component to also use the abstract popover so it conforms to the new API design, by doing so also fixes an issue with the grid using a destroyed view via nested ng-content.

closes #1116 #1215

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>