[NG/UI] Select 2.0
This is a first basic implementation of the Select 2.0 drafted here #248
it does only support single select and doesn't care for scrolling or anything.

Update#1: it now supports server-driven options with the same api known from the datagrid component namely (clrDgRefresh) [clrDgLoading]
and a addional one [clrLoadingText] to be able to customize the Loading Text the approach in the draft is possible but i'm not sure where to get by the label information

I'm posting this PR to get early feedback on it, it is very much not meant to be merged on its own as it need polishing and more tests.

Signed-off-by: Marvin Osswald <mail@marvinosswald.de>