[NG] Internal structural directive for virtual scrolling
CAUTION: This directive is for internal use only, we do not support its use outside of our own components.

See demos for code examples on how to use it.
It's still fairly crude and missing many customizations, but we can grow it organically as we go.

Whenever we are making a change to this, we'll need to check for potential performance issues and
make sure it runs correctly on all browsers.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>