[NG] fix bug with selection with pagination and server-driven datagrids
Since you cannot use clrDgItems with server-driven datagrids, we need to ensure
the trackBy function is leveraged as part of the comparison process for
determining if something is equal or not. This is due to server-driven datagrids
getting new objects with each page (from an API for example), and those don't
strictly equal one another. You must use trackBy for selection to display
correctly, though selection does persist without it. This also fixes an
unreported issue where clrDgItems did not handle trackBy that used the index
value.

closes #1393

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>