[NG] Tabs chocolate
PR #1514 introduced chocolate errors in tabs. So I reverted it and rewrote a new fix for #1318.

As now standard with chocolate errors, we let Willy Wonka and the Oompa Loompas handle it. The actual fix is 10 lines long, but I wanted to add unit tests for it and got sucked in when trying to improve the helpers slightly. So there is a bunch on non-related tests cleanups, all revolving around stronger, safer typing.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>