[CORE] Mention location of docs in CONTRIBUTING.md
[CORE] Mention location of documentation website in CONTRIBUTING.md

The documentation for clarity is provided by the https://vmware.github.io/clarity/documentation/ website.  This is stored in this repository under the `new-website` branch, which is not indexed by github, so searching for text snippets from the documentation will not return any results.  This adds a brief mention to `CONTRIGBUTING.md` which tells people were to look for the documentation if they need to propose changes. 

Signed-off-by: Alex Mellnik <a.r.mellnik@gmail.com>