[UI] deprecate the existing forms implementation
This moves all of the existing forms code into a new folder so a new
implementation can be started. It also renames the ClrCheckbox to
ClrCheckboxDeprecated, with a deprecation so it doesn't break but show show in
an IDE. This will help organize new forms work as a separate unit so we can
simply remove this when we need.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>