[NG] Preserve selection on items refresh for client-side datagrids
Fixes several bugs at once where we used the wrong variables (or didn't use them at all) to try and preserve selection, which meant we cleared it every time.
Also fixes the corresponding spec, which were completely broken for two reasons: a nested `beforeEach()` that didn't do anything, and the fact that we tested synchronously that an asynchronous operation didn't change anything. In other words, we were testing that code that hasn't run yet did not have any effects...

Fixes #1593.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>