[NG] Clean up datagrid footer's flexbox
The only child of the footer that now expands is the
`.datagrid-foot-description` which contains any custom text or buttons
the consumer decides to add to the footer. Since this element is always
here, it will naturally push its siblings to the left and to the right
of the footer, preserving the current styles.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>