[UI] Enhancement for tooltip (#1955)
Provide a SASS variable to control the borders of a tooltip (including the
arrow). In order to implement the border for the arrow, we implemented
using before, after pseudo elements. One triangle is created with the
background color and other triangle is created using border color. The
values are adjusted for all the variant of toolips
We have tested

Different variant Tooltips such as top-right, top-left, bottom right,
bottom-left, right and left
Different browsers such as Chrome, Mozilla, and Edge
Also tested signposts and dont see any dependencies
In order to adjust different tooltips in different browsers we have
adjusted the values as much as possible with best possible outcomes.
Please find attached the images for the same.

Fixes #1955

Signed-off-by: Uttam Agarwal <uttamagarwal.slg@gmail.com>