[CORE] Contribution process and coding guidelines
Title says it all. We're committing them to the repository for contributors to have access to them locally as soon as they pull the project, and it will also allow us to have discussions about this process when submitting PRs to update it.
The design contribution guideline is somewhere else as far as I know, I will see with Red when we can port it here.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>