[NG] Set isRootLevelToggle to public
When running a strict configuration of ngc (e.g. ng-packagr) against a library
which imports the ClrDropdownModule (e.g. ng-packagr), ngc errors out because
the dropdown's private property is used in the "host" component. Note that
even though the compiler appears to access clarity, the component isn't actually
bundled in the resulting output.

Signed-off-by: Michael Krotscheck <krotscheck@gmail.com>