[UI] fix build for 0.12 where relative paths cause issues for theming (#2336)
Using relative links to node_modules in our build works fine for our build, but if someone tries to build a theme and imports our file, the file paths are no longer correct. We had the tilde importer in the past, but it doesn't seem to work the same without additional webpack functionality. This will require anyone building a custom theme to add the `--include-path=node_modules` to their build as well.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>