[NG] Input working with grid automatically, and host wrapping on inputs
This adds a few new directives and a service to manage the layout option for the form. It will use that to conditionally add the new grid classes that will auto-flip to vertical on narrow screens regardless of layout choice. It also adds host wrapping to the input so it will always be able to layout correctly.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>