[NG] Proper Datagrid typing
This is a pass over our datagrid code to parametrize everything that needs to by the types of items we receive. It drastically reduces the number of `any` and improves the stability of our code.

I also removed old casts that are not needed anymore since both Angular and Typescript evolved and can now guess types better.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>