[NG] Internationalizing accessibility
This is a rough first draft to label every icon button and
every label-less checkbox and radio button we use in our
components with a very generic screenreader-only text.

Because this text has to be internationalized for the
components to be accessible in every language, we use a
provider to declare these labels. That way our consumers can
declare their own "CommonStringsService" to override ours with
the labels they want in any language. And their provider can
obviously inject the app's current locale with Angular so they
can fetch the correct strings dynamically.

Additional notes:
- For invalid form inputs, the actually error message should be
enough for screen readers. I added `aria-hidden="true"` to them
for now, but I'm open to discussion on this. One more involved
solution would be to use `aria-labelledby` on the exclamation
circle icon to label it with the actual error message. But that
requires trading `id`s and it's out of the scope of this PR.
- No unit tests have been implemented for these labels (except
the alert ones) because we want this to be part of a broader
a11y testing tool.
- Upgrade Typescript version for a cool feature, but realized
Angular 6 uses Typescript 3.0+ anyway.
- The CSS change on alerts is to fix a weird screenreader issue
due to the lack of display: block. No idea why.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>