[ENH] Allowing tooltips to be used inside static button groups
This will enable a workaround for people who want to use tooltips in their button groups. It is a fix specific to Firefox. To enable the angular tooltips in Firefox we would need to do a much more involved refactor of the API.

Tested in:
✔︎ Chrome
✔︎ Firefox

Signed-off-by: Scott Mathis <smathis@vmware.com>

PR changes

We shouldn't use our angular component tags in CSS.

Signed-off-by: Scott Mathis <smathis@vmware.com>

Updating contribution guidelines

Signed-off-by: Scott Mathis <smathis@vmware.com>

PR changes

Signed-off-by: Scott Mathis <smathis@vmware.com>