[BUILD] This adds a custom TSLint rule for us to guard against barrel imports
This has been a challenge to always catch correctly in reviews, so this will automatically throw linting errors (even in your IDE) if you are importing from a barrel instead of the file directly. This also fixes several instances of this issue that don't seem to cause build issues, because they are just interfaces and dropped from the final build anyways or are in spec files.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>