[NG] Use a `providedIn: 'root'` for our i18n service
Because we used a manual declaration in our main module, people
importing feature modules did not have access to this provider.
I initially did not use the new syntax because I believed it was
incompatible with factory providers, but it turns out it was
just undocumented.

This change introduces a new circular dependency warning between
two of our modules, but it seems that even the most basic use
of tree-shakeable factory providers will always produce a
circular dependency. Let's wait until the dust settles on the
new syntax to see what's the Angular-recommended way to solve
the circular dependency.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>