Dev app: import only feature modules in lazy-loaded modules
The correct way for real-life application to import Clarity in production is to only import the feature modules they need, in order to reduce the size of their bundle. Components declared in an NgModule cannot be treeshaken.

Our dev app did not follow this pattern, which means it wasn't representative of actual applications using Clarity. This change fixes that to have an accurate sandbox when developing.
Also note that the kitchen sink app still imports `ClarityModule`, so the main module is still tested as part of our suite.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>