[NG] Tree view: Lazy declarative case (#2838)
The nodes are declared in HTML on the app's side, but the app uses `*clrIfExpanded` to only instantiate them (and maybe fetch them) when needed. Because of this, we cannot propagate selection down so we simply notify the app when the user selects a node and let them figure out what it means for their model.
In other words, "lazy" is the equivalent of server-driven datagrid: if we have imperfect information on the dataset, we just do not try to be smart and become a dumb UI, the logic is on the app's side or on their server.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>