[NG] Tree view: Recursive case, both eager and lazy
In the recursive case, our consumers give us the root nodes of the tree
and a function to get the children of a node. We use this to build the model
of the tree, and we then create the node components (or to be more precise,
the templates that contain the node components) based on that pure model.

The way to do this is to use a custom structural directive to iterate over
the root nodes, and the children-fetching function follows the same API pattern
as the TrackByFunction from *ngFor.

Once again, in the eager case we build everything on initialization and try to
be smart about selection, whereas in the lazy case we take inputs literally and
let the app or server handle the consequences of selecting a node.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>