[NG] Models for the new tree (#2813)
Just introducing the models for now. They share the selection logic and propagation, and the specifics of each are that:
- The declarative model is pretty dumb since it just waits for the components to create it. It just registers/unregisters to the parent to make sure our structure is up to date.
- The recursive model is smarter and allows us to build the whole tree in memory (by calling `fetchChildren()` recursively, hence the name) even if we don't have the corresponding components. We will use that model to instantiate the node components ourselves.

Also note that I'm abandoning the double, tightly-couple, binding for both `selected` and `indeterminate`. It created way too many issues both for us and for the apps. I'm using a single binding with a three-values enum, which ends up being much more natural to use and cuts down a lot of the internal logic.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>