[WEB] Fix wizard overflow on the website (#2828)
I am piercing shadow DOM with ::ng-deep, which is not recommended generally,
but I'm using it exceptionally to fix an issue, that will only be available
on the website. I am protecting this change not to spread to other components
on the site by preceding it with a :host, which makes the CSS more specific.
Tested on Chrome Canary, Chrome normal, Safari, FF, Edge, IE11.
Change merged to 1.0, 0.12, 0.13, as the issue is not observed on older versions.

Signed-off-by: Ivan Donchev <idonchev@vmware.com>