[NG] New tree implementation (#2859)
Implementation was reviewed in the following PRs:

1. Models for the new tree (#2813)

Just introducing the models for now. They share the selection logic and propagation, and the specifics of each are that:
- The declarative model is pretty dumb since it just waits for the components to create it. It just registers/unregisters to the parent to make sure our structure is up to date.
- The recursive model is smarter and allows us to build the whole tree in memory (by calling `fetchChildren()` recursively, hence the name) even if we don't have the corresponding components. We will use that model to instantiate the node components ourselves.

Also note that I'm abandoning the double, tightly-couple, binding for both `selected` and `indeterminate`. It created way too many issues both for us and for the apps. I'm using a single binding with a three-values enum, which ends up being much more natural to use and cuts down a lot of the internal logic.

2. Tree view: Eager declarative case (#2827)

This case is the simple one: the nodes are declared in HTML on the app's side, and we know all of them from the start so we can be smart about selection.

3. Tree view: Lazy declarative case (#2838)

The nodes are declared in HTML on the app's side, but the app uses `*clrIfExpanded` to only instantiate them (and maybe fetch them) when needed. Because of this, we cannot propagate selection down so we simply notify the app when the user selects a node and let them figure out what it means for their model.
In other words, "lazy" is the equivalent of server-driven datagrid: if we have imperfect information on the dataset, we just do not try to be smart and become a dumb UI, the logic is on the app's side or on their server.

4. Tree view: Recursive case, both eager and lazy (#2845)

In the recursive case, our consumers give us the root nodes of the tree
and a function to get the children of a node. We use this to build the model
of the tree, and we then create the node components (or to be more precise,
the templates that contain the node components) based on that pure model.

The way to do this is to use a custom structural directive to iterate over
the root nodes, and the children-fetching function follows the same API pattern
as the TrackByFunction from *ngFor.

Once again, in the eager case we build everything on initialization and try to
be smart about selection, whereas in the lazy case we take inputs literally and
let the app or server handle the consequences of selecting a node.

5. Tree view: Asynchronous cases (#2851)

This commit adds the integration of our loading directive to the tree, to display a nice spinner instead of a caret when waiting for the children to be fetched. Works for both the declarative and recursive cases.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>