[ICONS] automatically include core icons into Angular modules, build SVG files into a separate directory so they aren't shipped with @clr/icons package on NPM, and support binding to [shape] property on clr-icon.
In order to better serve users, we will handle loading the core icon sets in the icon module, since this is imported anytime a module needs to use icons. We can improve this with loading specific icons as we go later on.

This also cleans up some internals and makes things more consistent for our naming conventions and guidelines.

closes #2353

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>