[WEBSITE] adding the website to the main project
This is a whole sale move of the files from the existing website branch over to master, which will allow us to do docs and features at the same time. It also includes the basics for netlify to use as a deployment platform, fixes up all linting and build issues, and is in sync with latest website.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>