[NG] feat(smart-popover) This is the first PR for adding smart popovers that position themselves inteligently and without any clipping issues. There are several things I'm adding here.
There is a stub for ClrIfOpen b/c I Started to modify ClrIfOpen withthe `put elements on body` approach but realized that it will affect all of the popovers that already use *clrIfOpen. Doing it that way would mean *all* popovers need to be switched to the SmartPopover before this could be completed. The second page has a datagrid with a hidable column on it For now I'm using it to develop the SmartPopover functionality but will also use it for documentation when its ready for that.

- ClrIfOpen: started as a place I could put popover components that were already using it. Needs better description and explanation to showcase the utilty.
- ClrSmartOpen: What I'm calling the new directive that is pulling the template out of the view and attaching it to the `document.body`

If you think this resembles the ClrIfOpen directive a little bit you are correct. There are a couple of differences that affect how I wrote the tests. It also did have a (much smaller) waterfall affect on the ClrDatagridColumnToggle component unit tests. Those changes are also reflected here.

This is a stub class that will get filled out with the event logic (scrolling, clicking and ignoring clicks) in a future PR.
Signed-off-by: Matt Hippely <mhippely@vmware.com>

Signed-off-by: Matt Hippely <mhippely@vmware.com>