feat(toggle) new toggle implementation for form controls
This component is largely just a piggy back off of the checkbox component because toggles are checkboxes with a different visual treatment. That is why the checkbox and toggle selectors are bundled together and there is a service to track if the checkbox or toggle was defined, and display accordingly.

closes #2953

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>