[NG] Stop caching children for lazy recursive trees
We used to fetch children for lazy recursive trees only once, but we closed their selection Observable when we shouldn't. Rather than avoid closing the Observables when we reuse models, I decided to stop caching children altogether, letting the app decide whether they want to cache or not and do it on their side. It's less intrusive and more flexible.

Fixes #3209.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>