fix(tree) set tree view node event emitter to synchronous and don't output selected changes when setting input on node
A tree view with a number of nodes could be tricked into a state of uncertainty (flickering) by double clicking fast to trigger too many async event loops in the bindings. This stops the event propagation from firing in this specific case, but keeps it in the remainder where it is needed to update other nodes.

fixes #3073

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>