[NG] Keyboard accessible dropdown
This commit uses the common focus management utilities to enable keyboard navigation of dropdowns.

The biggest hurdle here is that inside of a dropdown menu, the focusable items can be either a dropdown item or the trigger of a nested dropdown. But because we can't query deeply as that would break multiple levels of nesting, our `@ContentChildren` query only has access to the top-level children. This means we have to work with the nested dropdowns themselves, not their trigger. To get this to work, I introduce a custom FocusableItem implementation for each dropdown which "proxies" focus behavior to its trigger, and bridges keyboard navigation by pointing to its children.

Note that this relies on the fact that `@ContentChildren` doesn't have to query for components or directives, it can query directly for providers. So any component or directive declaring said provider will be part of the resulting `QueryList`.

This implementation uses a pattern you might not have faced yet with Observables: observables that execute an action when any observer subscribes to them (in this case opening/closing the dropdown). This lets us avoid adding a bunch of observables in the FocusService to try and subscribe to them from every focusable item, which would end up in nice spaghetti code. Instead, when the focus service tries to go down, we can simply intercept that with our on-subscribe pattern, open the dropdown, and then go down.

There were a lot of edge cases to test, I hope I my specs are comprehensive enough.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>