Metropolis font metrics `backtick` fix
Regression found on Unicode character 0060 width allocation and fixed

New metrics revised and a 303-glyph set supplied as a replacement for the inline Clarity CSS.




Font packaging sidenote:
Highly recommend: removing all inlined BASE 64 chunks. Not only does this add to the overload payload size of Clarity, but it is perhaps inefficient because a text format is used to carry not one, but multiple fonts. Newer font encoding methodologies have since been released, such as WOFF2 for modern browsers. I understand Issue #3269 was recently made, hope it gets in soon. As for dealing with FOUT issue, if and when it occurs, I can recommend clean workarounds in the browser.

Signed-off-by: Matt blacknugget@excite.com