chore(core): Update packages for Angular v8.0.0 support
As part of upgrading for Clarity v2.0.0 `ng update @angular/cli @angular/core` was run. Changes to config and code files were made to address failing testing and build errors. ted:

1. Angular packages were updated
2. Gemini is commented out: Angular 8 requires node 12 and a gemini dependency doesn't compile on node 12
3. Added node-sass back: for some reason sass (sass-dart was not getting picked up and build errors out)
4. Modified test utils to assign functions directly (Object.assign is only assigning properties after the update)
5. Updates jasmine packages
6. Updated misc package files that were behind
7. Application routes were migrated to the new lazy module form
8. Updates ViewChild (s) in website and fixes the website app

`ng update` made changes to the tsconfig files (module: esnext) that changes the way functions are imported and spied on when there is no parental object. I removed those changes but want to call them out for attention during the review since this seems to be something the cli is now defaulting to.

There is one `xit` in the wizard header action tests. Ti is failing and it looks like the DOM isn't fully cleaned up in between tests. Still need to investigate more here.

Signed-off-by: Matt Hippely <mhippely@vmware.com>