Remove datagrid animation hack (#3339)
The hack is replaced with Angular animation. The hack was used because
dynamic animations were not possible in old Angular versions.
The cleanup change for the focus trap is needed because of testing
issues with NoopAnimationsModule. Though it only affects tests, we'd
better keep the tests and make minor modification to the code, as with
the current NoopAnimationsModule implementation we can't fix the tests
other ways.

Signed-off-by: Ivan Donchev <idonchev@vmware.com>