fix(website) remove the need to use SSR
This turns off the SSR and Angular universal configuration, as it appears Google Search is smart enough now not to need prerendered pages. SEO was the driver for adding it to begin with, and this fixes the build for the website by not needing SSR instead of waiting for updates to packages. This also removes compodoc, which is having fits, until we have a better strategy. Also removes node-sass because its not needed.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>