fix(popover): renamed files and entities
This change adds the popover utility entities to the ClrPopoverNext namespace.
This included the following name changes to the entities in the utility codebase:

1. Removed ClrSmart from all service entities
2. Remove  from directives, selectors and where appropriate I replaced it with
3. Files were renamed according to our code conventions and to align with the entity name changes.

NOTE: These changes are only for file and entity name changes. All code functionality should be the same as the last commit to this branch.
This prepares the code base for rebasing and merge with master.

Signed-off-by: Matt Hippely <mhippely@vmware.com>