refactor(i18n): replace original provider with new implementation
BREAKING CHANGE: The previous i18n service would get removed from a production build in Angular since it wasn't
referenced. To address this, there is now a service that Clarity provides itself and you can pass a custom,
new localized set of strings into the service instead of trying to use the provider tree in Angular.
This will be helpful to future clr-base.

Signed-off-by: Jeremy Wilken <gnomation@gnomeontherun.com>