build: supress output from visual test script
The docker kill command was logging output to the command line when trying to kill an instance of clarity_chrome that wasn't already running.
Initially I think I put this there to prevent issues restarting but its changed over time. Because we are moving away from gemin I'm supressing the output of the kill command so its not logged rather than re-working the script logic.

Signed-off-by: Matt Hippely <mhippely@vmware.com>