refactor(datagrid): move row controls into sticky container
This moves the datagrid row controls into the stick container and prepares for pinned columns (frontside).
Styles were adjusted to handle the column separaters and flex the sticky and scrollable elements in the
header as a row instead of a column.

An additional visual change was made to add a small sperator at the end of the sticky cell so that its more obvious that the scrollable cells are sliding underneath the sticky content.

RELATED CHANGES:
The pokemon generator was modified to produce consistent items from the three arrays that are used to
generate data so that visual regression tests could be turned on and utilized for visual regression testing
of the datagrid use cases. This commit will generate reference images for the baseline.

Signed-off-by: Matt Hippely <mhippely@vmware.com>